Refactored the AssetType
