+++ src/backend/PegasusAssets.cpp
@@ -17,7 +17,7 @@
 
 #include "PegasusAssets.h"
 
-#include "AssetTypes.h"
+#include "types/AssetType.h"
 #include "utils/HashMap.h"
 
 #include <QString>
