+++ src/backend/types/types.pri
@@ -1,5 +1,6 @@
 HEADERS += \
     $$PWD/AppCloseType.h \
+    $$PWD/AssetType.h \
     $$PWD/KeyEventType.h \
     $$PWD/ProviderType.h \
     $$PWD/GamepadKeyId.h \
